# docs/swagger.yaml
definitions:
  flightticket.FlightTicket:
    properties:
      arrival:
        type: integer
      category:
        $ref: '#/definitions/flightticket.FlightTicketCategories'
      city_from:
        type: string
      city_to:
        type: string
      id:
        type: string
      quantity:
        type: integer
      take_off:
        type: integer
      value:
        type: integer
    type: object
  flightticket.FlightTicketCategories:
    enum:
    - самый быстрый
    - самый дешевый
    - самый дешевый, самый быстрый
    - ""
    type: string
    x-enum-varnames:
    - Fastest
    - Cheapest
    - CheapestFastest
    - None
...
  httputils.ErrorResponse:
    properties:
      error:
        example: failed to parse query value 'stars-from'
        type: string
      message:
        example: failed to find hotels
        type: string
      success:
        example: false
        type: boolean
    type: object
  httputils.SuccessResponse:
    properties:
      data: {}
      message:
        example: successfully found hotels
        type: string
      success:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
paths:
  /api/admin/flight-ticket/create:
    post:
      consumes:
      - application/json
      description: Создаёт авиабилет с переданными параметрами
      parameters:
      - description: Из какого города
        in: query
        name: city-from
        required: true
        type: string
      - description: В какой город
        in: query
        name: city-to
        required: true
        type: string
      - description: Количество билетов/мест
        in: query
        name: quantity
        required: true
        type: integer
      - description: Цена за билет
        in: query
        name: value
        required: true
        type: integer
      - description: 'Дата и время взлета (формат: 2006-01-02T15:04:05Z)'
        in: query
        name: take-off
        required: true
        type: string
      - description: 'Дата и время посадки (формат: 2006-01-02T15:04:06Z)'
        in: query
        name: arrival
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httputils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/flightticket.FlightTicket'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      summary: Создать авиабилет (Admin only)
      tags:
      - Авиабилет
  /api/admin/flight-ticket/delete:
    delete:
      consumes:
      - application/json
      description: Удаляет авиабилет по ID
      parameters:
      - description: ID авиабилета
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputils.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      summary: Удалить авиабилет (Admin only)
      tags:
      - Авиабилет
  /api/admin/flight-ticket/update:
    put:
      consumes:
      - application/json
      description: Изменяет авиабилет переданными параметрами
      parameters:
      - description: ID авиабилета
        in: query
        name: id
        required: true
        type: string
      - description: Новый из какого города (необязатено)
        in: query
        name: city-from
        type: string
      - description: Новый в какой город (необязатено)
        in: query
        name: city-to
        type: string
      - description: Новое количество билетов/мест (необязатено)
        in: query
        name: quantity
        type: integer
      - description: Новая цена за билет (необязатено)
        in: query
        name: value
        type: integer
      - description: 'Новые дата и время взлета (формат: 2006-01-02T15:04:05Z) (необязатено)'
        in: query
        name: take-off
        type: string
      - description: 'Новые дата и время посадки (формат: 2006-01-02T15:04:06Z) (необязатено)'
        in: query
        name: arrival
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputils.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      summary: Изменить авиабилет (Admin only)
      tags:
      - Авиабилет
...
  /api/flight-ticket/buy:
    post:
      consumes:
      - application/json
      description: Покупает авиабилеты по ID и количеству пассажиров. Требуется авторизация
      parameters:
      - description: ID авиабилета
        in: query
        name: id
        required: true
        type: string
      - description: Количество пассажиров
        in: query
        name: quantity
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httputils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/flightticket.FlightTicket'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      summary: Купить авиабилет
      tags:
      - Авиабилет
  /api/flight-ticket/find:
    get:
      consumes:
      - application/json
      description: Находит авиабилет по фильтрам. Можно искать без фильтров. Можно
        использовать 'city-from' и 'city-to' вместе с 'city-via' для поиска билетов
        с пересадкой в этом городе или без него. Если указать 'arrange', то билеты
        будут отсортированы по цене в указанном порядке (asc - по возрастанию, desc
        - по убыванию). Самый быстрый и самый дешевый авиабилеты/рейсы отмечаются
        категориями 'самый быстрый', 'самый дешевый', исключение если нет прямого
        пути без пересадок - в таком случае выдается самый быстрый путь
      parameters:
      - description: ID авиабилета (необязатено)
        in: query
        name: id
        type: string
      - description: Из какого города (необязатено)
        in: query
        name: city-from
        type: string
      - description: Через какой город (необязатено)
        in: query
        name: city-via
        type: string
      - description: В какой город (необязатено)
        in: query
        name: city-to
        type: string
      - description: Количество билетов/мест (необязатено)
        in: query
        name: quantity
        type: integer
      - description: Цена за билет (необязатено)
        in: query
        name: value
        type: integer
      - description: 'Дата и время взлета (формат: 2006-01-02T15:04:05Z) (необязатено)'
        in: query
        name: take-off
        type: string
      - description: 'Дата и время посадки (формат: 2006-01-02T15:04:06Z) (необязатено)'
        in: query
        name: arrival
        type: string
      - description: Упорядочить по цене ('asc' возрастание, 'desc' убывание) (необязатено)
        in: query
        name: arrange
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputils.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.ErrorResponse'
      summary: Найти авиабилет
      tags:
      - Авиабилет
...
swagger: "2.0"